pyproject.toml
[tool.poetry]
name = "course-4"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.7.0"
psycopg2 = "^2.9.10"
gunicorn = "^23.0.0"
django-debug-toolbar = "^5.0.1"
django-extensions = "^3.2.3"
celery = "^5.4.0"
django-allauth = "^65.4.1"
dotenv = "^0.9.9"
send-mail = "^0.0.2"
pillow = "^11.1.0"
psycopg2-binary = "^2.9.10"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
